% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adversarialbehavior.R
\docType{data}
\name{adversarialbehavior}
\alias{adversarialbehavior}
\title{adversarialbehavior Bayesian Network}
\format{
A discrete Bayesian network to define and solve the inference problem of adversarial movement in the grid infrastructure towards targets of physical impact. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{RemoteAdversary}{(TRUE, FALSE);}
\item{RootAccessFTPServer}{(TRUE, FALSE);}
\item{FTPBasedBufferOverflow}{(TRUE, FALSE);}
\item{RemoteBufferOverflowOnSSHDaemon}{(TRUE, FALSE);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Inferring adversarial behaviour in cyber-physical power systems using a Bayesian attack graph approach.
}
\references{
Sahu, A., & Davis, K. (2023). Inferring adversarial behaviour in cyber-physical power systems using a Bayesian attack graph approach. IET Cyber-Physical Systems: Theory & Applications, 8(2), 91-108.
}
