% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agropastoral2.R
\docType{data}
\name{agropastoral2}
\alias{agropastoral2}
\title{agropastoral Bayesian Networks}
\format{
A discrete Bayesian network to explore  the relationship between the environment and social organisation (Fig. 6 of the paper). The structure of the BN was given within the referenced paper together with a dataset. Probabilities were learned using the dataset and the discretization mentioned in the paper. The vertices are:
\describe{
\item{Annual_Temperature}{(low, medium, high);}
\item{Community_Size}{(<200, >=200)}
\item{CV_Annual_Precipitation}{(low, medium, high);}
\item{CV_Annual_Productivity}{(low, medium, high);}
\item{CV_Annual_Temperature}{(low, medium, high);}
\item{Distance_Coast}{(low, medium, high);}
\item{Elevation}{(low, medium, high);}
\item{Landscape}{(Aquatic, Tundra, Desert, Forest, Grassland);}
\item{Monthly_Precipitation}{(low, medium, high);}
\item{Monthly_Productivity}{(low, medium, high);}
\item{Slope}{(low, medium, high);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Exploring the role of ecology and social organisation in agropastoral societies: A Bayesian network approach.
}
\references{
Palacios, O., Barceló, J. A., & Delgado, R. (2022). Exploring the role of ecology and social organisation in agropastoral societies: A Bayesian network approach. Plos One, 17(10), e0276088.
}
