% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algalactivity1.R
\docType{data}
\name{algalactivity1}
\alias{algalactivity1}
\title{algalactivity Bayesian Networks}
\format{
A discrete Bayesian network to to predict chlorophyll-a (chl-a) using a range of water quality parameters as predictors (Fig. 6 of the referenced paper). Probabilities were given within the referenced paper (a uniform was given to the vertex Chl_a since it was missing). The vertices are:
\describe{
\item{C}{(0, 1);}
\item{Chl_a}{(0, 1);}
\item{DO}{(0, 1);}
\item{N}{(0, 1);}
\item{P}{(0, 1);}
\item{pH}{(0, 1);}
\item{Te}{(0, 1);}
\item{Tu}{(0, 1);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Influence of resampling techniques on Bayesian network performance in predicting increased algal activity.
}
\references{
Rezaabad, M. Z., Lacey, H., Marshall, L., & Johnson, F. (2023). Influence of resampling techniques on Bayesian network performance in predicting increased algal activity. Water Research, 244, 120558.
}
