% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms1.R
\docType{data}
\name{algorithms1}
\alias{algorithms1}
\title{algorithms Bayesian Networks}
\format{
A Gaussian Bayesian network to illustrate the algorithms developed in the associated paper (Figure 1, top). The probabilities were available from a repository. The vertices are:
\describe{
\item{X1}{}
\item{X2}{}
\item{X3}{}
\item{X4}{}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Entropy and the Kullback-Leibler divergence for Bayesian networks: Computational complexity and efficient implementation.
}
\references{
Scutari, M. (2024). Entropy and the Kullback-Leibler Divergence for Bayesian Networks: Computational Complexity and Efficient Implementation. Algorithms, 17(1), 24.
}
\keyword{GBN}
