% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms4.R
\docType{data}
\name{algorithms4}
\alias{algorithms4}
\title{algorithms Bayesian Networks}
\format{
A discrete Bayesian network to illustrate the algorithms developed in the associated paper (Figure 2, bottom). The probabilities were available from a repository. The vertices are:
\describe{
\item{X1}{(a, b);}
\item{X2}{(c, d);}
\item{X3}{(e, f);}
\item{X4}{(g, h);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Entropy and the Kullback-Leibler divergence for Bayesian networks: Computational complexity and efficient implementation.
}
\references{
Scutari, M. (2024). Entropy and the Kullback-Leibler Divergence for Bayesian Networks: Computational Complexity and Efficient Implementation. Algorithms, 17(1), 24.
}
