% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beatles.R
\docType{data}
\name{beatles}
\alias{beatles}
\title{Beatles Bayesian Network}
\format{
A discrete Bayesian modelling a situation where a member of the Beatles band might be dead. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{ExactlyOneBeatlesIsDead}{(TRUE, FALSE);}
\item{GeorgeIsAlive}{(TRUE, FALSE);}
\item{JohnIsAlive}{(TRUE, FALSE);}
\item{PaulIsAlive}{(TRUE, FALSE);}
\item{RingoIsAlive}{(TRUE, FALSE);}
}
}
\description{
Measuring coherence with Bayesian networks
}
\references{
Kowalewska, A., & Urbaniak, R. (2023). Measuring coherence with Bayesian networks. Artificial Intelligence and Law, 31(2), 369-395.
}
\keyword{ReverseNaiveBayes}
