% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnRep.R
\name{bnRep}
\alias{bnRep}
\title{bnRep: A Repository of Bayesian Network Models}
\description{
A repository of discrete, Gaussian, and conditional linear Gaussian Bayesian networks from the recent academic literature.
}
\details{
The package includes over 200 Bayesian networks which appeared in recent academic papers. They can be accessed by their name, as provided in this documentation.

They are stored as \code{bn.fit} objects from the \code{bnlearn} package. Recall that in order to plot them, the function \code{bn.net} must be used to convert them into a graph object.

The package includes two handy functionalities:

\itemize{
\item The \code{bnRep_summary} object: a dataframe including a lot of details about the Bayesian networks in the repository;
\item The \code{bnRep_app} function, which launchs a Shiny app to explore the Bayesian networks in the repository.
}

Thanks to the interface with \code{bnlearn}, functions from that package can be used to export the networks in other formats and use them in other platforms, such as Netica, Hugin, or Python.
}
