% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnRep_app.R
\name{bnRep_app}
\alias{bnRep_app}
\title{Launch the Bayesian Network Viewer App}
\usage{
bnRep_app()
}
\description{
This function launches the Shiny app that allows users to interactively view and filter the Bayesian networks repository.
}
