% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burglar.R
\docType{data}
\name{burglar}
\alias{burglar}
\title{burglar Bayesian Network}
\format{
A discrete Bayesian network modeling a simple burglary scenario (Model 1, Table 2). The network was available from an associated repository. The vertices are:
\describe{
\item{Burglar}{(Suspect 1, Suspect 2, Suspect3);}
\item{PrimaryItemStolen}{(Jewellery, Electronics, Money);}
\item{BurglaryTime}{(Day, Night);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Strategies for selecting and evaluating information.
}
\references{
Liefgreen, A., Pilditch, T., & Lagnado, D. (2020). Strategies for selecting and evaluating information. Cognitive Psychology, 123, 101332.
}
\keyword{NaiveBayes}
