% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachexia1.R
\docType{data}
\name{cachexia1}
\alias{cachexia1}
\title{cachexia Bayesian Networks}
\format{
A Gaussian Bayesian networks comparing the dependence of metabolomics for people who suffer of Cachexia. The Bayesian network is learned as in the referenced paper. The vertices are:
\describe{
\item{A}{Adipate;}
\item{B}{Betaine;}
\item{F}{Fumarate;}
\item{GC}{Glucose;}
\item{GM}{Glutamine;}
\item{V}{Valine;}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Model-preserving sensitivity analysis for families of Gaussian distributions.
}
\references{
Gorgen, C., & Leonelli, M. (2020). Model-preserving sensitivity analysis for families of Gaussian distributions. Journal of Machine Learning Research, 21(84), 1-32.
}
\keyword{GBN}
