% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.R
\docType{data}
\name{case}
\alias{case}
\title{case Bayesian Network}
\format{
A discrete Bayesian network for concrete legal fact idioms that qualify events in a narrative Bayesian network. The network was available from a public repository. The vertices are:
\describe{
\item{Body}{(f, t);}
\item{ComplicityMurder}{(f, t);}
\item{DebtFightFK}{(f, t);}
\item{FBarn}{(f, t);}
\item{FightBarn}{(f, t);}
\item{FStenGun}{(f, t);}
\item{Help}{(f, t);}
\item{Intent}{(f, t);}
\item{KBarn}{(f, t);}
\item{Killed}{(f, t);}
\item{KKilled}{(f, t);}
\item{Murder}{(f, t);}
\item{Murdered}{(f, t);}
\item{PlanBarnF}{(f, t);}
\item{Premed}{(f, t);}
\item{Prov}{(f, t);}
\item{SBarn}{(f, t);}
\item{ShootStenGun}{(none, F, Not F);}
\item{TMathus}{(f, t);}
\item{TSF1}{(f, t);}
\item{TSF2}{(f, t);}
\item{TSLocation}{(f, t);}
\item{TStenGun}{(f, t);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Building a stronger case: Combining evidence and law in scenario-based Bayesian networks.
}
\references{
Van Leeuwen, L., Verbrugge, R., Verheij, B., & Renooij, S. (2024, June). Building a Stronger Case: Combining Evidence and Law in Scenario-Based Bayesian Networks. In 3rd International Conference on Hybrid Human-Artificial Intelligence, HHAI 2024 (pp. 291-299). IOS Press.
}
