% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concrete4.R
\docType{data}
\name{concrete4}
\alias{concrete4}
\title{concrete Bayesian Networks}
\format{
A discrete Bayesian network for evaluating the presence probability of blue concrete (Fig. E1 - Single-Family Houses, of the referenced paper). Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{AverageDistance}{(0-300, 300-600, >600);}
\item{Basement}{(False, True);}
\item{BlueConcrete}{(False, True);}
\item{ConstructionYear}{(1930-1955, 1955-1960, 1960-1968, 1968-1975, 1975-1980);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Estimating the probability distributions of radioactive concrete in the building stock using Bayesian networks.
}
\references{
Wu, P. Y., Johansson, T., Mangold, M., Sandels, C., & Mjornell, K. (2023). Estimating the probability distributions of radioactive concrete in the building stock using Bayesian networks. Expert Systems with Applications, 222, 119812.
}
\keyword{Tree}
