% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concrete5.R
\docType{data}
\name{concrete5}
\alias{concrete5}
\title{concrete Bayesian Networks}
\format{
A discrete Bayesian network for evaluating the presence probability of blue concrete (Fig. E1 - Multi-Family Houses, of the referenced paper). Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{AverageDistance}{(0-300, 300-600, >600);}
\item{BlueConcrete}{(False, True);}
\item{FloorArea}{(0-150, 150-220, 220-360, 360-1500, >1500);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Estimating the probability distributions of radioactive concrete in the building stock using Bayesian networks.
}
\references{
Wu, P. Y., Johansson, T., Mangold, M., Sandels, C., & Mjornell, K. (2023). Estimating the probability distributions of radioactive concrete in the building stock using Bayesian networks. Expert Systems with Applications, 222, 119812.
}
\keyword{Tree}
