% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concrete6.R
\docType{data}
\name{concrete6}
\alias{concrete6}
\title{Concrete Bayesian Networks}
\format{
A discrete Bayesian network for evaluating the presence probability of blue concrete (Fig. E1 - School Buildings, of the referenced paper). Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{AverageDistance}{(0-300, 300-600, >600);}
\item{BlueConcrete}{(False, True);}
\item{FloorArea}{(0-150, 150-220, 220-360, 360-1500, >1500);}
}
}
\description{
Estimating the probability distributions of radioactive concrete in the building stock using Bayesian networks.
}
\references{
Wu, P. Y., Johansson, T., Mangold, M., Sandels, C., & Mjornell, K. (2023). Estimating the probability distributions of radioactive concrete in the building stock using Bayesian networks. Expert Systems with Applications, 222, 119812.
}
\keyword{DBN}
\keyword{NaiveBayes}
