% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concrete7.R
\docType{data}
\name{concrete7}
\alias{concrete7}
\title{Concrete Bayesian Networks}
\format{
A discrete Bayesian network for evaluating the presence probability of blue concrete (Fig. E1 - Other Buildings, of the referenced paper). Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{AverageDistance}{(0-300, 300-600, >600);}
\item{BlueConcrete}{(False, True);}
\item{ConstructionYear}{(1930-1955, 1955-1960, 1960-1968, 1968-1975, 1975-1980);}
\item{FloorArea}{(0-150, 150-220, 220-360, 360-1500, >1500);}
\item{NumberOfStairwells}{(0, 1, 2, 3, 4);}
}
}
\description{
Estimating the probability distributions of radioactive concrete in the building stock using Bayesian networks.
}
\references{
Wu, P. Y., Johansson, T., Mangold, M., Sandels, C., & Mjörnell, K. (2023). Estimating the probability distributions of radioactive concrete in the building stock using Bayesian networks. Expert Systems with Applications, 222, 119812.
}
\keyword{DBN}
