% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructionproductivity.R
\docType{data}
\name{constructionproductivity}
\alias{constructionproductivity}
\title{constructionproductivity Bayesian Network}
\format{
A discrete Bayesian network to identify causal relationship and occurrence probability of critical factors affecting construction productivity. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{Accidents}{(Yes, No);}
\item{AdverseWeather}{(Yes, No);}
\item{Age}{(Yes, No);}
\item{Attitude}{(Yes, No);}
\item{EngineerQualification}{(Yes, No);}
\item{Experience}{(Yes, No);}
\item{HealthStatus}{(Yes, No);}
\item{MaterialPresence}{(Yes, No);}
\item{OwnerFinance}{(Yes, No);}
\item{PlanningAndMethod}{(Yes, No);}
\item{Productivity}{(Yes, No);}
\item{Sex}{(Yes, No);}
\item{SkilledWorkers}{(Yes, No);}
\item{TaskComplexity}{(Yes, No);}
\item{TechnologyLevel}{(Yes, No);}
\item{WorkingFrequency}{(Yes, No);}
\item{WorkingTools}{(Yes, No);}
\item{Workmanship}{(Yes, No);}
}
}
\description{
Construction productivity prediction through Bayesian networks for building projects: case from Vietnam.
}
\references{
Khanh, H. D., & Kim, S. Y. (2022). Construction productivity prediction through Bayesian networks for building projects: Case from Vietnam. Engineering, Construction and Architectural Management, 30(5), 2075-2100.
}
