% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corticosteroid.R
\docType{data}
\name{corticosteroid}
\alias{corticosteroid}
\title{corticosteroid Bayesian Network}
\format{
A discrete Bayesian network to compute the conditional probability of complete clinical response and remission. The probabilities were given within the referenced paper. The vertices are:
\describe{
\item{FinalCSDCAchieved}{(Achieved, Not Achieved);}
\item{CCRAchieved}{(Achieved, Not Achieved);}
\item{RemissionAchieved}{(Achieved, Not Achieved);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Corticosteroid discontinuation, complete clinical response and remission in juvenile dermatomyositis.
}
\references{
Kishi, T., Warren-Hicks, W., Bayat, N., Targoff, I. N., Huber, A. M., Ward, M. M., ... & with the Childhood Myositis Heterogeneity Study Group. (2021). Corticosteroid discontinuation, complete clinical response and remission in juvenile dermatomyositis. Rheumatology, 60(5), 2134-2145.
}
