% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\docType{data}
\name{crypto}
\alias{crypto}
\title{crypto Bayesian Network}
\format{
A discrete Bayesian modelling to exam- ine the causal interrelationships among six major cryptocurrencies. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{Bitcoin}{(Down, Up);}
\item{Binance_Coin}{(Down, Up);}
\item{Ethereum}{(Down, Up);}
\item{Tether}{(Down, Up);}
\item{Litecoin}{(Down, Up);}
\item{Ripple}{(Down, Up);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Dynamic evolution of causal relationships among cryptocurrencies: an analysis via Bayesian networks.
}
\references{
Amirzadeh, R., Thiruvady, D., Nazari, A., & Ee, M. S. (2024). Dynamic evolution of causal relationships among cryptocurrencies: an analysis via Bayesian networks. Knowledge and Information Systems, 1-16.
}
