% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drainage.R
\docType{data}
\name{drainage}
\alias{drainage}
\title{drainage Bayesian Network}
\format{
A discrete Bayesian network for fault diagnosis of a coal mine drainage system. The probabilities were available from a repository. The vertices are:
\describe{
\item{AbnormalFlow}{(T, F);}
\item{AirLeakageOfShaftSeal}{(T, F);}
\item{GetValveFailure}{(T, F);}
\item{ImpellerDamaged}{(T, F);}
\item{LowSpeed}{(T, F);}
\item{LowVoltage}{(T, F);}
\item{MotorFault}{(T, F);}
\item{PipelineFailure}{(T, F);}
\item{PipelineRupture}{(T, F);}
\item{Silting}{(T, F);}
\item{WaterPumpFailure}{(T, F);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Fuzzy Bayesian network fault diagnosis method based on fault tree for coal mine drainage system.
}
\references{
Shi, X., Gu, H., & Yao, B. (2024). Fuzzy Bayesian Network Fault Diagnosis Method Based on Fault Tree for Coal Mine Drainage System. IEEE Sensors Journal.
}
\keyword{ReverseTree}
