% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ets.R
\docType{data}
\name{ets}
\alias{ets}
\title{ets Bayesian Network}
\format{
A discrete Bayesian network to model the influence of financial, economic, and energy-related factors on EUA futures prices. The model was given in the referenced paper. The vertices are:
\describe{
\item{CAC}{(High, Low, Neutral);}
\item{CO1}{(High, Low, Neutral);}
\item{DAX}{(High, Low, Neutral);}
\item{ECO}{(High, Low, Neutral);}
\item{EURCHF}{(High, Low, Neutral);}
\item{EURCNY}{(High, Low, Neutral);}
\item{EURGBP}{(High, Low, Neutral);}
\item{EURRUB}{(High, Low, Neutral);}
\item{EURUSD}{(High, Low, Neutral);}
\item{LBEATREU}{(High, Low, Neutral);}
\item{LB01TREU}{(High, Low, Neutral);}
\item{MO1}{(High, Low, Neutral);}
\item{MXEU0EN}{(High, Low, Neutral);}
\item{NG1COMB}{(High, Low, Neutral);}
\item{SPGTCED}{(High, Low, Neutral);}
\item{SPX}{(High, Low, Neutral);}
\item{SXXP}{(High, Low, Neutral);}
\item{VIX}{(High, Low, Neutral);}
\item{XA1}{(High, Low, Neutral);}
\item{XAU}{(High, Low, Neutral);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Uncovering drivers of EU carbon futures with Bayesian networks.
}
\references{
Maciejowski, J., & Leonelli, M. (2025). Uncovering Drivers of EU Carbon Futures with Bayesian Networks. arXiv preprint arXiv:2505.10384.
}
