% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firealarm.R
\docType{data}
\name{firealarm}
\alias{firealarm}
\title{firealarm Bayesian Network}
\format{
A discrete Bayesian network to model a simple fire alarm system. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{Fire}{(true, false);}
\item{Tampering}{(true, false);}
\item{Smoke}{(true, false);}
\item{Alarm}{(true, false);}
\item{Leaving}{(true, false);}
\item{Report}{(true, false);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
When do numbers really matter?.
}
\references{
Hei Chan, Adnan Darwiche (2002). "When do numbers really matter?". Journal of Artificial Intelligence Research 17 (265-287).
}
