% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluids2.R
\docType{data}
\name{fluids2}
\alias{fluids2}
\title{fluids Bayesian Networks}
\format{
A discrete Bayesian network to assess the presence of saliva in the recovered material and combine potentially contradictory observations. The network was available from an associated repository. The vertices are:
\describe{
\item{Risk}{Risk of false positive for saliva detection (High, Low);}
\item{Saliva}{(Yes, No);}
\item{RSID}{Saliva test (Positive, Negative, Weak positive);}
\item{Concentration}{Concentration of total DNA (0-0.0002, 0.0002-0.0005, 0.0005-0.001, 0.001-0.002, 0.002-0.004, 0.004-0.01, 0.01-0.01, 0.02-inf);}
\item{Nature_of_stain}{(Saliva, Fecal matter/vaginal secretion/sperm/breat milk/urine, Other);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Use of Bayesian Networks for the investigation of the nature of biological material in casework.
}
\references{
Samie, L., Champod, C., Delemont, S., Basset, P., Hicks, T., & Castella, V. (2022). Use of Bayesian Networks for the investigation of the nature of biological material in casework. Forensic Science International, 331, 111174.
}
\keyword{ReverseTree}
