% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluids3.R
\docType{data}
\name{fluids3}
\alias{fluids3}
\title{fluids Bayesian Networks}
\format{
A discrete Bayesian network to assess the presence of sperm in the recovered material and combine potentially contradictory observations. The network was available from an associated repository. The vertices are:
\describe{
\item{Concentration_EPI}{Total concentration of male DNA in non sperm fraction (0-0.0002, 0.0002-0.0005, 0.0005-0.001, 0.001-0.002, 0.002-0.004, 0.004-0.01, 0.01-0.01, 0.02-inf);}
\item{Sperm}{(Yes, No);}
\item{Nature_of_stain}{(At least Sperm, Lubricant/urine/vaginal secretion);}
\item{Location}{(Vaginal/condom/panties, Other);}
\item{Concentration_Total}{Total concentration of male DNA (0-0.0002, 0.0002-0.0005, 0.0005-0.001, 0.001-0.002, 0.002-0.004, 0.004-0.01, 0.01-0.01, 0.02-inf);}
\item{AZO}{(Azoospermic, Non azoospermic);}
\item{CT}{Spermatozoa detection (Positive, Negative, 1 spz, Possible spz);}
\item{PSA}{Seminal fluid test (Positive, Negative, Weak positive);}
\item{Concentration_SP}{Total concentration of male DNA in sperm fraction (0-0.0002, 0.0002-0.0005, 0.0005-0.001, 0.001-0.002, 0.002-0.004, 0.004-0.01, 0.01-0.01, 0.02-inf);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Use of Bayesian Networks for the investigation of the nature of biological material in casework.
}
\references{
Samie, L., Champod, C., Delemont, S., Basset, P., Hicks, T., & Castella, V. (2022). Use of Bayesian Networks for the investigation of the nature of biological material in casework. Forensic Science International, 331, 111174.
}
