% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodsecurity.R
\docType{data}
\name{foodsecurity}
\alias{foodsecurity}
\title{Foodsecurity Bayesian Network}
\format{
A discrete Bayesian network modelling a food security scenario. Probabilities were given within the referenced paper. The vertices are Cost, Educational_Attainment, Health, and Social_Cohesion.
}
\description{
Coherent combination of probabilistic outputs for group decision making: an algebraic approach.
}
\references{
Leonelli, M., Riccomagno, E., & Smith, J. Q. (2020). Coherent combination of probabilistic outputs for group decision making: an algebraic approach. OR Spectrum, 42(2), 499-528.
}
\keyword{GBN}
