% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_list.R
\name{get_network_list}
\alias{get_network_list}
\title{Get the list of available Bayesian network files}
\usage{
get_network_list()
}
\value{
A character vector of network file names.
}
\description{
This function lists all the .rda files in the data directory.
}
