% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grounding.R
\docType{data}
\name{grounding}
\alias{grounding}
\title{grounding Bayesian Network}
\format{
A discrete Bayesian network to for quantitative analysis of the causation of grounding accidents in Arctic shipping. Probabilities were given within the referenced paper (some information appeared incorrect). The vertices are:
\describe{
\item{BW}{Bad Weather (No,Yes);}
\item{DAM}{Damage (No,Yes);}
\item{DE}{Dependent equipment (No,Yes);}
\item{GRO}{Grounding (No,Yes);}
\item{ICC}{Insufficient communication and collaboration (No,Yes);}
\item{IER}{Imperfect emergency (No,Yes);}
\item{ILC}{Improper labeling of the chart (No,Yes);}
\item{INE}{Inefficient use of navigation equipment (No,Yes);}
\item{IO}{Improper operation (No,Yes);}
\item{IPS}{Insufficient preparation for sailing (No,Yes);}
\item{IRP}{Improper route planning (No,Yes);}
\item{IRR}{Irregularities (No,Yes);}
\item{IS}{Insufficient supervision (No,Yes);}
\item{ISL}{Inconsistent standardization and language (No,Yes);}
\item{ISS}{Insufficient supervision system, rules and regulations (No,Yes);}
\item{IWP}{Insufficient work plan (No,Yes);}
\item{LID}{Limited information dissemination channels (No,Yes);}
\item{LNE}{Lack of navigation equipment (No,Yes);}
\item{LSM}{Lack of safety management system (No,Yes);}
\item{LT}{Lack of training (No,Yes);}
\item{MIJ}{Misjudgment (No,Yes);}
\item{OD}{Outdated data (No,Yes);}
\item{PC}{Poor communication at high latitudes (No,Yes);}
\item{PEC}{Poor external communication (No,Yes);}
\item{PF}{Psychological factors (No,Yes);}
\item{PFC}{Poor traffic conditions (No,Yes);}
\item{PSA}{Poor situational awareness (No,Yes);}
\item{PSM}{Poor safety management (No,Yes);}
\item{PSQ}{Poor service quality (No,Yes);}
\item{SMS}{Ship SMS conflict (No,Yes);}
\item{UCD}{Unupdated chart data (No,Yes);}
\item{UDL}{Unclear division of labour (No,Yes);}
\item{UPA}{Unreasonable planning and arrangement (No,Yes);}
\item{UR}{Underestimate the risk (No,Yes);}
\item{US}{Unsafe speed (No,Yes);}
\item{WD}{Wrong decision (No,Yes);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
A framework for quantitative analysis of the causation of grounding accidents in arctic shipping.
}
\references{
Fu, S., Yu, Y., Chen, J., Xi, Y., & Zhang, M. (2022). A framework for quantitative analysis of the causation of grounding accidents in arctic shipping. Reliability Engineering & System Safety, 226, 108706.
}
