% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydraulicsystem.R
\docType{data}
\name{hydraulicsystem}
\alias{hydraulicsystem}
\title{hydraulicsystem Bayesian Network}
\format{
A discrete Bayesian network to to analyze time series hydraulic system operation reliability. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{YellowHydraulicSystem}{(True, False);}
\item{GreenHydraulicSystem}{(True, False);}
\item{BlueHydraulicSystem}{(True, False);}
\item{HydraulicSystem}{(True, False);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Analysis and assessment of risks to public safety from unmanned aerial vehicles using fault tree analysis and Bayesian network.
}
\references{
Pan, W. H., Feng, Y. W., Liu, J., & Lu, C. (2023). Operation reliability monitoring towards fault diagnosis of airplane hydraulic system using quick access recorder flight data. Measurement Science and Technology, 34(5), 055111.
}
\keyword{ReverseNaiveBayes}
