% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liquefaction.R
\docType{data}
\name{liquefaction}
\alias{liquefaction}
\title{liquefaction Bayesian Network}
\format{
A Gaussian Bayesian network to predict seismic liquefaction-induced settlement. The Bayesian network is learned using the data available from the referenced paper. The vertices are:
\describe{
\item{Ds}{}
\item{GWT}{}
\item{lnamax}{}
\item{lnR}{}
\item{lnt}{}
\item{Mw}{}
\item{N160}{}
\item{S}{}
\item{Sigmav}{}
\item{Ts}{}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
A continuous Bayesian network regression model for estimating seismic liquefaction-induced settlement of the free-field ground.
}
\references{
Hu, J., Xiong, B., Zhang, Z., & Wang, J. (2023). A continuous Bayesian network regression model for estimating seismic liquefaction-induced settlement of the free-field ground. Earthquake Engineering & Structural Dynamics, 52(11), 3216-3237.
}
\keyword{GBN}
