% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_name.R
\name{load_network}
\alias{load_network}
\title{Load a Bayesian network}
\usage{
load_network(network_name)
}
\arguments{
\item{network_name}{The name of the network file to load.}
}
\value{
A bn.fit object representing the Bayesian network.
}
\description{
This function loads a selected Bayesian network file.
}
