% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\docType{data}
\name{navigation}
\alias{navigation}
\title{navigation Bayesian Network}
\format{
A discrete Bayesian network to demonstrate the existing variability in functions that are part of the complex navigation system. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{ToAcquireGPSsignal}{(Accurate, Acceptable, Inaccurate);}
\item{ToCheckAircraftPositionExecutingRNPProcedure}{(Accurate, Acceptable, Inaccurate);}
\item{ToKeepAircraftOnProgrammedRoute}{(Accurate, Acceptable, Inaccurate);}
\item{ToShowAircraftPositionBasedOnGPSSignal}{(Accurate, Acceptable, Inaccurate);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Safety analysis of RNP approach procedure using fusion of FRAM model and Bayesian belief network.
}
\references{
Oliveira, D., Moraes, A., Junior, M. C., & Marini-Pereira, L. (2023). Safety analysis of RNP approach procedure using fusion of FRAM model and Bayesian belief network. The Journal of Navigation, 76(2-3), 286-315.
}
\keyword{ReverseTree}
