% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perioperative.R
\docType{data}
\name{perioperative}
\alias{perioperative}
\title{perioperative Bayesian Network}
\format{
A discrete Bayesian network for the estimation of the drug suspension period even in the presence of competing risks. The probabilities were available from a repository. The vertices are:
\describe{
\item{DrugSuspension}{(0 days, 5 days, 7 days);}
\item{ThromboticRisk}{(High, Medium, Low);}
\item{BleedingRisk}{(High, Null);}
\item{PlateletCount}{(High, Medium, Low);}
\item{AbnormalAPTT}{(High, Medium, Low);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Development of a perioperative medication suspension decision algorithm based on Bayesian networks.
}
\references{
Kawaguchi, S., Fukuda, O., Kimura, S., Yeoh, W. L., Yamaguchi, N., & Okumura, H. (2024, January). Development of a Perioperative Medication Suspension Decision Algorithm Based on Bayesian Networks. In 2024 IEEE/SICE International Symposium on System Integration (SII) (pp. 7-12). IEEE.
}
\keyword{NaiveBayes}
