% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot.R
\docType{data}
\name{pilot}
\alias{pilot}
\title{pilot Bayesian Network}
\format{
A discrete Bayesian network to classify ADHD symptom. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{AdverseSeaSwell}{(Yes, No);}
\item{AdverseWind}{(Yes, No);}
\item{CommercialPressure}{(Yes, No);}
\item{ExcessiveEnvironmentFactors}{(Yes, No);}
\item{ExcessiveMotionVessel}{(Yes, No);}
\item{ExcessiveShipSpeed}{(Yes, No);}
\item{FailureHandholds}{(Yes, No);}
\item{HeavyRain}{(Yes, No);}
\item{HumanFailures}{(Yes, No);}
\item{ImproperShipHandling}{(Yes, No);}
\item{InappropriateAngle}{(Yes, No);}
\item{IncorrectHeigth}{(Yes, No);}
\item{IncorrectRigging}{(Yes, No);}
\item{IndividualFailure}{(Yes, No);}
\item{LackOfSafetyCulture}{(Yes, No);}
\item{LackOfSupervision}{(Yes, No);}
\item{ManeouveringFailures}{(Yes, No);}
\item{NonCertifiedPilotLadder}{(Yes, No);}
\item{NonComplyTrapdoor}{(Yes, No);}
\item{OperationalFailures}{(Yes, No);}
\item{OrganizationalFailure}{(Yes, No);}
\item{PilotLadder}{(Yes, No);}
\item{PilotTransferAccident}{(Yes, No);}
\item{PoorCombinationLadder}{(Yes, No);}
\item{PoorCommunicationWithPilotBoat}{(Yes, No);}
\item{PoorConditionPTA}{(Yes, No);}
\item{PoorIllumination}{(Yes, No);}
\item{PoorISMSystem}{(Yes, No);}
\item{PoorPilotLadder}{(Yes, No);}
\item{PTAEquipmentFailure}{(Yes, No);}
\item{PTAFailure}{(Yes, No);}
\item{PTAPreparedWindward}{(Yes, No);}
\item{RestrictedVisibility}{(Yes, No);}
\item{RetrievalLine}{(Yes, No);}
\item{RiggingFailure}{(Yes, No);}
\item{SecuringFailure}{(Yes, No);}
\item{SecuringFailurePilot}{(Yes, No);}
\item{SecuringFailurePTA}{(Yes, No);}
\item{ShipSideObstructed}{(Yes, No);}
\item{StructuralFailure}{(Yes, No);}
\item{SubstandardActs}{(Yes, No);}
\item{SubstandardConditions}{(Yes, No);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Dynamic analysis of pilot transfer accidents.
}
\references{
Sakar, C., & Sokukcu, M. (2023). Dynamic analysis of pilot transfer accidents. Ocean Engineering, 287, 115823.
}
\keyword{ReverseTree}
