% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rockburst.R
\docType{data}
\name{rockburst}
\alias{rockburst}
\title{rockburst Bayesian Network}
\format{
A Gaussian Bayesian network to give early-warning of rockbursts. The probabilities were given within the referenced paper. The vertices are:
\describe{
\item{Rockburst}{(No, Yes);}
\item{MMAV}{(Sligth, Medium, Strong);}
\item{SRAV}{(Small, Medium, Big);}
\item{ASAV}{(Small, Medium, Big);}
\item{DSDAV}{(Small, Medium, Big);}
\item{SEAV}{(Low, Medium, High);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Dynamic early warning of rockburst using microseismic multi-parameters based on Bayesian network.
}
\references{
Li, X., Mao, H., Li, B., & Xu, N. (2021). Dynamic early warning of rockburst using microseismic multi-parameters based on Bayesian network. Engineering Science and Technology, an International Journal, 24(3), 715-727.
}
\keyword{TAN}
