% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sallyclark.R
\docType{data}
\name{sallyclark}
\alias{sallyclark}
\title{sallyclark Bayesian Network}
\format{
A discrete Bayesian modelling the evidence from the Sally Clark trial. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{ABrusing}{(TRUE, FALSE);}
\item{ADisease}{(TRUE, FALSE);}
\item{AMurder}{(TRUE, FALSE);}
\item{BBruising}{(TRUE, FALSE);}
\item{BDisease}{(TRUE, FALSE);}
\item{BMurder}{(TRUE, FALSE);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Measuring coherence with Bayesian networks.
}
\references{
Kowalewska, A., & Urbaniak, R. (2023). Measuring coherence with Bayesian networks. Artificial Intelligence and Law, 31(2), 369-395.
}
\keyword{Tree}
