% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonella1.R
\docType{data}
\name{salmonella1}
\alias{salmonella1}
\title{salmonella Bayesian Networks}
\format{
A discrete Bayesian network to show the existence of dependencies between resistance to antimicrobials. Probabilities were given within the referenced paper. The vertices are (s stands for susceptible, r for resistant):
\describe{
\item{CHL}{Chloramphenicol (s, r);}
\item{CIP}{Ciprofloxacin (s, r);}
\item{CTX}{Cefotaxime (s, r);}
\item{FFC}{Florfenicol (s, r);}
\item{GEN}{Gentamicin (s, r);}
\item{NAL}{Nalidixic acid (s, r);}
\item{TET}{Tetracycline (s, r);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Patterns of antimicrobial resistance in Salmonella isolates from fattening pigs in Spain.
}
\references{
Teng, K. T. Y., Aerts, M., Jaspers, S., Ugarte-Ruiz, M., Moreno, M. A., Saez, J. L., ... & Alvarez, J. (2022). Patterns of antimicrobial resistance in Salmonella isolates from fattening pigs in Spain. BMC Veterinary Research, 18(1), 333.
}
