% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tastingtea.R
\docType{data}
\name{tastingtea}
\alias{tastingtea}
\title{tastingtea Bayesian Network}
\format{
A discrete Bayesian network for modelling the Lady Tasting Tea experiment. The probabilities were given in the referenced paper. The vertices are:
\describe{
\item{AbilityToTaste}{(0.5, 0.75, 1);}
\item{Cup1}{(tea, milk);}
\item{Cup2}{(tea, milk);}
\item{Cup3}{(tea, milk);}
\item{Cup4}{(tea, milk);}
\item{Cup5}{(tea, milk);}
\item{Cup6}{(tea, milk);}
\item{Cup7}{(tea, milk);}
\item{Cup8}{(tea, milk);}
\item{TestOutcome1}{(tea, milk);}
\item{TestOutcome2}{(tea, milk);}
\item{TestOutcome3}{(tea, milk);}
\item{TestOutcome4}{(tea, milk);}
\item{TestOutcome5}{(tea, milk);}
\item{TestOutcome6}{(tea, milk);}
\item{TestOutcome7}{(tea, milk);}
\item{TestOutcome8}{(tea, milk);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
A Bayesian network for modelling the Lady tasting tea experiment.
}
\references{
Xie, G. (2024). A Bayesian network for modelling the Lady tasting tea experiment. PloS one, 19(7), e0307866.
}
