% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theft2.R
\docType{data}
\name{theft2}
\alias{theft2}
\title{theft Bayesian Networks}
\format{
A discrete Bayesian network representing a legal scenario. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{AtCrimeScene}{(F, T);}
\item{EredHanded}{(F, T);}
\item{EseenCS}{(F, T);}
\item{EWallet}{(F, T);}
\item{Guilty}{(F, T);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Evaluating methods for setting a prior probability of guilt.
}
\references{
van Leeuwen, L., Verheij, B., Verbrugge, R., & Renooij, S. (2023). Evaluating Methods for Setting a Prior Probability of Guilt. In Legal Knowledge and Information Systems (pp. 63-72). IOS Press.
}
\keyword{Tree}
