% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaccine.R
\docType{data}
\name{vaccine}
\alias{vaccine}
\title{Vaccine Bayesian Network}
\format{
A (synthetic) discrete Bayesian network modeling a vaccine scenario. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{Screening_Test}{(Negative, Positive);}
\item{Disease}{(Healthy, Mildly, Severly);}
\item{Vaccine}{(No, Yes);}
}
}
\description{
Sensitivity analysis in multilinear probabilistic models.
}
\references{
Leonelli, M., Gorgen, C., & Smith, J. Q. (2017). Sensitivity analysis in multilinear probabilistic models. Information Sciences, 411, 84-97.
}
\keyword{DBN}
