% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agropastoral5.R
\docType{data}
\name{agropastoral5}
\alias{agropastoral5}
\title{agropastoral Bayesian Networks}
\format{
A discrete Bayesian network to explore the relationship between social organisation and subsistence strategies. The structure of the BN was given within the referenced paper together with a dataset. Probabilities were learned using the dataset and the discretization mentioned in the paper. The vertices are:
\describe{
\item{Community_Organisation}{(Clan communities, Missing, No exogamous clans);}
\item{Community_Size}{(<200, >=200);}
\item{Gathering}{(None, <25, >=25);}
\item{Household_Organisation}{(Small extended, Large extended, Nuclear);}
\item{Settlement_Types}{(Camp, Hamlet, Homesteads, Village);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Exploring the role of ecology and social organisation in agropastoral societies: A Bayesian network approach.
}
\references{
Palacios, O., Barceló, J. A., & Delgado, R. (2022). Exploring the role of ecology and social organisation in agropastoral societies: A Bayesian network approach. Plos One, 17(10), e0276088.
}
\keyword{NaiveBayes}
