% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnRep_app.R
\name{bnRep_app}
\alias{bnRep_app}
\title{Launch the Bayesian Network Viewer App}
\usage{
bnRep_app()
}
\value{
The function calls a Shiny app to plot networks in \code{bnRep} and explore the database of networks stored in \code{bnRep_summary}.
}
\description{
This function launches the Shiny app that allows users to interactively view and filter the Bayesian networks repository.
}
