% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaltest.R
\docType{data}
\name{medicaltest}
\alias{medicaltest}
\title{medicaltest Bayesian Network}
\format{
A discrete Bayesian network representing a synthethic example of two medical tests. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{Test1}{(no, yes);}
\item{Test2}{(no, yes);}
\item{Disease}{(no, yes);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Global sensitivity analysis of uncertain parameters in Bayesian networks.
}
\references{
Ballester-Ripoll, R., & Leonelli, M. (2024). Global Sensitivity Analysis of Uncertain Parameters in Bayesian Networks. arXiv preprint arXiv:2406.05764.
}
\keyword{ReverseNaiveBayes}
