% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{simulation Bayesian Network}
\format{
A discrete Bayesian network to assist asset managers in evaluating the risk arising from the operations. Probabilities were given within the referenced paper. The vertices are:
\describe{
\item{JointFailure}{(True, False);}
\item{PressureRegulatorLeakage}{(True, False);}
\item{SealFailure}{(True, False);}
\item{ValveActivation}{(True, False);}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
Integration of fuzzy reliability analysis and consequence simulation to conduct risk assessment.
}
\references{
Leoni, L., & De Carlo, F. (2023). Integration of fuzzy reliability analysis and consequence simulation to conduct risk assessment. Journal of Loss Prevention in the Process Industries, 83, 105081.
}
\keyword{ReverseNaiveBayes}
