% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softwarelogs1.R
\docType{data}
\name{softwarelogs1}
\alias{softwarelogs1}
\title{softwarelogs Bayesian Networks}
\format{
A discrete Bayesian network to discover poor performance indicators in a system and to explore usage patterns that usually require temporal analysis. The networks are given in the referenced paper. The vertices are:
\describe{
\item{error}{Error that has occurred (com.mysql.jdbc.jdbc2.optional.MysqlDataSource, java.lang.IllegalArgumentException, java.lang.NullPointerException, java.lang.NumberFormatException, javax.ejb.EJBException, javax.ejb.EJBTransactionRolledbackException);}
\item{class}{Class that throws the error (chessleague.db.AccountFacade, chessleague.ejb.AccountBean, ... , org.jboss.weld.servlet.WeldTerminalListener);}
\item{method}{Method where the error has occured (_jspService, \if{html}{\out{<init>}}, deleteAccount, findClass, forInputString, getDataSourceObject, getInfo, getSessionContext, mapLocal3xException, processSystemException, publishMessage, sendHeadlines, undoBid);}
\item{severity}{Severity of the entry (SEVERE, WARNING, INFO);}
\item{thread_name}{Method where the error has occurred (AutoDeployer, ContainerBackgroundProcessor/ChessLeague-war, ContainerBackgroundProcessor, deployment-jar-scanner, http-listener, RunLevelControllerThread).}
}
}
\description{
Bayesian Network analysis of software logs for data‐driven software maintenance.
}
\references{
del Rey, S., Martinez-Fernandez, S., & Salmeron, A. (2023). Bayesian Network analysis of software logs for data-driven software maintenance. IET Software, 17(3), 268-286.
}
