% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{titanic Bayesian Network}
\format{
A discrete Bayesian network modeling the survival of the Titanic passengers. The Bayesian network was learned as in the referenced paper. The vertices are:
\describe{
\item{Class}{(1st, 2nd, 3rd, Crew);}
\item{Sex}{(Male, Female);}
\item{Age}{(Child, Adult);}
\item{Survived}{(No, Yes).}
}
}
\value{
An object of class \code{bn.fit}. Refer to the documentation of \code{bnlearn} for details.
}
\description{
The R Package stagedtrees for Structural Learning of Stratified Staged Trees.
}
\references{
Carli, F., Leonelli, M., Riccomagno, E., & Varando, G. (2022). The R Package stagedtrees for Structural Learning of Stratified Staged Trees. Journal of Statistical Software, 102, 1-30.
}
